ui_agent/requirements.txt
browser-use
openai
python-dotenv
requests
fastapi
uvicorn
